Optional repair update
Changed the calculation for repair amount more based on number of crew, skill of crew, and number of repair drones. Repairs happen less often, they are lower than vanilla with small amount of service crew who are low skilled. Repair drones are now a multiplier effect on the ability of service crew.

New repair rate is a minimum of 0.1% and a max of 5% hull and occurs around every 60%. Stations have a lower repair rate due to massive amount of HP and reasoning that they have less "combat" personnel onboard.